move some scoring utils to scoring module
