+++ app/routes/stats.$postId.tsx
@@ -7,9 +7,9 @@ import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/db.ts'
 import { getStatsPost } from '#app/modules/posts/post-repository.ts'
 import { getEffects } from '#app/modules/scoring/effect-repository.ts'
+import { relativeEntropy } from '#app/modules/scoring/scoring-utils.ts'
 import { type StatsPost } from '#app/types/api-types.ts'
 import { type DBEffect } from '#app/types/db-types.ts'
-import { relativeEntropy } from '#app/utils/entropy.ts'
 
 const postIdSchema = z.coerce.number()
